Break out origin and probing sphere scenes
